Basic Edge and Node infrastructure is in!
The Edge and Node classes offer simple creation and memory maintenance.  Destructed `Node`s will notify their `Edge`s so that the end can be set to `nullptr`.  If both tail and head of an `Edge` are `nullptr`, the `Edge` deletes itself.

Cookbook code in `test.cpp/h` shows usage of the `connector` and some user `Node` types with custom edge traversals.  Some of the functionality in there might get ported into the main library.